chore(reservation-application): 테스트 패키지 구조 정렬
